requirements.txt
# requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
anthropic>=0.7.0
openai>=1.0.0
speechrecognition>=3.10.0
pydub>=0.25.1
streamlit-webrtc>=0.47.0
python-dotenv>=1.0.0
plotly>=5.17.0
pyaudio>=0.2.11